Refactor home. Fixes and improves
