+++ app/src/app/app.component.scss
@@ -285,6 +285,10 @@
     padding: 15px 0px 0px 0px;
 }
 
+.main-item-div {
+    flex: 1 1 auto;
+  }
+
 .progress-bar {
     padding: 0px 16px 0px 16px;
 }
