+++ app/src/app/core/services/index.ts
@@ -9,3 +9,4 @@ export * from './control.service';
 export * from './calendar.service';
 export * from './settings.service';
 export * from './theme.service';
+export * from './home.service';
