+++ app/src/app/pages/home/home.page.scss
@@ -15,10 +15,6 @@
   margin: 0px;
 }
 
-.main-item-div {
-  flex: 1 1 auto;
-}
-
 .main-avatar {
   padding: 20px 0px 0px 14px;
 }
@@ -46,7 +42,7 @@
 }
 
 .icon-row2 {
-  padding: 5px 2px 8px 5px;
+  padding: 3px 2px 8px 5px;
 }
 
 .icon-row3 {
@@ -54,7 +50,7 @@
 }
 
 .h-row2 {
-  font-size: 16px;
+  //font-size: 16px;
   margin: 2px 2px 1px 1px;
 }
 
@@ -64,7 +60,7 @@
 }
 
 .p-row2 {
-  margin: 7px 8px 0px 1px;
+  margin: 5px 8px 0px 1px;
   color: var(--ion-color-step-600,#666);
   font-size: 14px;
 }
