+++ app/src/app/shared/modals/info-calendar/info-calendar.component.html
@@ -38,9 +38,9 @@
         <ion-icon *ngIf="!hideVehicles[i]" size="small" color="primary" name="caret-up-circle"></ion-icon>
       </ion-button>
     </ion-list-header>
-    <ion-item *ngFor="let maint of vehicle.listInfoCalendarMaintenance" [hidden]="hideVehicles[i]" button>
+    <ion-item *ngFor="let maint of vehicle.listInfoCalendarMaintenance" [hidden]="hideVehicles[i]">
       <!-- Icons -->
-      <ion-avatar slot="start" class="max-width10 align-start">
+      <ion-avatar slot="start" class="max-width9 align-start">
         <ion-icon size="large" class="icon-color" [name]="this.getIconMaintenance(maint)"></ion-icon>
       </ion-avatar>
       <div class="main-item">
@@ -59,18 +59,25 @@
                 color="icon-color-gray" class="icon-color-gray font-size16"></ion-icon>
             </div>
           </div>
-          <div *ngFor="let repl of maint.listInfoCalendarReplacement" (click)="showInfo(repl)">
-            <div class="table-row margin-right5 table-row auto-width">
-              <ion-icon [class]="getClassIcon(repl.warning, 'icon-pos')" [color]="getClassIcon(repl.warning, '')" size="small"
-                [name]="getIconKms(repl.warning)"></ion-icon>
-              <p class="p-color row-pos auto-width row-overflow-hiden font-size14">{{ repl.nameReplacement }}</p>
-              <p class="p-color row-pos align-right font-size14">{{ repl.dateFormat }}</p>
+          <ion-item button *ngFor="let repl of maint.listInfoCalendarReplacement" class="second-item" (click)="showInfo(repl)">
+            <!-- Ivatar -->
+            <ion-avatar slot="start" class="max-width9 align-start">
+              <ion-icon [class]="getClassIcon(repl.warning, 'second-icon-avatar')"
+                [color]="getClassIcon(repl.warning, '')" [name]="getIconReplacement(repl)"></ion-icon>
+            </ion-avatar>
+            <div class="main-item-div">
+              <div class="table-row margin-right5 table-row auto-width">
+                <ion-icon [class]="getClassIcon(repl.warning, 'icon-pos')" [color]="getClassIcon(repl.warning, '')" size="small"
+                  [name]="getIconKms(repl.warning)"></ion-icon>
+                <p class="p-color row-pos auto-width row-overflow-hiden font-size14">{{ repl.nameReplacement }}</p>
+                <p class="p-color row-pos align-right font-size14">{{ repl.dateFormat }}</p>
+              </div>
+              <div class="table-row margin-right5 table-row auto-width">
+                <p *ngIf="repl.time === 0 && repl.km !== 0" class="subrow-color subrow-pos align-right auto-width subrow-width font-size13 font-italic">{{ 'COMMON.DISTANCE' | translate }}: {{ repl.km }}{{measure.value}}</p>
+                <p *ngIf="repl.price !== 0" class="subrow-color subrow-pos align-right auto-width font-size13 font-italic">{{ 'COMMON.LAST_PRICE' | translate }}: {{ repl.price }}{{coin.value}}</p>
+              </div>
             </div>
-            <div class="table-row margin-right5 table-row auto-width">
-              <p *ngIf="repl.time === 0 && repl.km !== 0" class="subrow-color subrow-pos align-right auto-width subrow-width font-size13 font-italic">{{ 'COMMON.DISTANCE' | translate }}: {{ repl.km }}{{measure.value}}</p>
-              <p *ngIf="repl.price !== 0" class="subrow-color subrow-pos align-right auto-width font-size13 font-italic">{{ 'COMMON.LAST_PRICE' | translate }}: {{ repl.price }}{{coin.value}}</p>
-            </div>
-          </div>
+          </ion-item>
         </div>
       </div>
     </ion-item>
