+++ app/src/app/shared/modals/info-calendar/info-calendar.component.scss
@@ -78,7 +78,7 @@
     margin: 0;
     padding-top: 1px;
     padding-bottom: 6px;
-    padding-left: 5px;
+    padding-left: 7px;
 }
 
 .font-size20 {
@@ -97,8 +97,22 @@
     font-size: 13px;
 }
 
+.second-icon-avatar {
+    padding-top: 7px;
+    padding-left: 1px;
+    font-size: 25px;
+}
+
 .main-item {
     width: 100%;
-    padding: 8px 0px 4px 0px;
+    padding: 8px 0px 0px 0px;
+}
+
+.second-item {
+    margin-right: -20px;
+}
+
+.max-width9 {
+    width: 9%;
 }
