+++ app/src/app/shared/modals/info-notification/info-notification.component.scss
@@ -4,4 +4,5 @@
 
 .margin-chart {
     margin-inline-start: -12px;
+    padding-bottom: 5px;
 }
